move UnreadSourceColumn to query's file
